generate new addresses when game restarts
