--- src/utils/GameState.ts
@@ -1,1 +1,0 @@
-export type GameState = "PLAYING" | "ACCESS_GRANTED" | "LOCKED_OUT"
