+++ src/utils/setUpGame.ts
@@ -10,32 +10,3 @@ export const wordLength = 7
 export const numCols = 2
 export const numLines = 16
 export const symbolsPerLine = symbolArrayLength/(numCols*numLines)
-
-export const addresses = generateAddresses()
-
-
-
-function generateAddresses(){
-    const byteSize = 8; //TODO: define bytesize elsewhere?
-    let addresses = Array(numLines*numCols)
-
-    const startingAddress = generateStartingAddress();
-
-    for(let i=0; i<addresses.length; i++){
-        addresses[i] = startingAddress+(i*byteSize);
-    }
-
-    return addresses;
-}
-
-function generateStartingAddress(){
-    const byteSize = 8;
-    const maxAddress = 0xFFFF - (numLines*numCols*byteSize)+byteSize;
-
-    let startingAddress = Math.ceil(Math.random()*maxAddress);
-
-    const remainder = startingAddress%byteSize;
-    startingAddress = startingAddress - remainder; //Removing the remainder ensures the address is a multiple of 8
-
-    return startingAddress;
-}
